tss enrichment proofing
