+++ episcanpy/tools/__init__.py
@@ -22,3 +22,5 @@ from ..tools._comparisons import transfer_obs, imputation
 
 from ..preprocessing._decomposition import tfidf, lsi
 from ..preprocessing._decomposition import nmf, fa
+
+from ..preprocessing._tss_enrichment import filter_enrichment_score
