+++ requirements.txt
@@ -12,7 +12,7 @@ networkx
 natsort
 joblib
 numba>=0.50.0
-numpy>=1.17.0
+numpy>=1.21.2
 bamnostic>=1.0.12
 umap-learn>=0.3.10
 tqdm
